fix: agent-backend communication and ticket lifecycle
- Add X-API-Key authentication support (verifyTokenOrAgent middleware)
- Fix route ordering in tickets.js (agent routes before parameterized routes)
- Fix TicketService static/instance method mismatch
- Add executePut() to ApiService for ticket status updates
- Update agent URL paths and authentication headers
- Add agent Dockerfile debug variant
- Expand integration test coverage

+++ agent/src/main/java/com/vibecode/agent/config/AgentConfig.java
@@ -68,7 +68,7 @@ public class AgentConfig {
         return backendUrl.endsWith("/") ? backendUrl + "api" : backendUrl + "/api";
     }
 
-    public String getGitHubBranchName(int ticketId, String ticketTitle) {
+    public String getGitHubBranchName(Long ticketId, String ticketTitle) {
         String slug = ticketTitle.toLowerCase()
             .replaceAll("[^a-z0-9]+", "-")
             .replaceAll("^-|-$", "")
